Warning fix.

+++ InteliHubClient/tioclient.c
@@ -1615,7 +1615,7 @@ clean_up_and_return:
 
 unsigned long get_n_readable_bytes(SOCKET sock) 
 {
-	unsigned long n = -1;
+	unsigned long n = (unsigned long)(-1);
 	if (ioctlsocket(sock, FIONREAD, &n) < 0) 
 	{
 		/* look in WSAGetLastError() for the error code */
